Add wallet section component
